requirements.txt

Markdown==3.3.6
notebook==6.4.11
numpy==1.22.3
pandas==1.4.2
psycopg2==2.8.6
python-dateutil==2.8.2
python-dotenv==0.20.0
spotipy==2.19.0
SQLAlchemy==1.3.24
streamlit==1.9.0
virtualenv==20.14.1

